Move code to identify MOSMIX-files to be read to separate file
