+++ lib/forecast_utils.js
@@ -3,115 +3,14 @@
 
 'use strict'
 
-const csv = require('dwd-csv-helper')
-const moment = require('moment')
-const path = require('path')
 const _ = require('lodash')
 const gu = require('../lib/general_utils.js')
-const bunyan = require('bunyan')
-const log = bunyan.createLogger({
-  name: 'dwd-csv-handler',
-  serializers: bunyan.stdSerializers
-})
 
+const bmu = require('./beob_mosmix_utils.js')
 const cd2u = require('./cosmo-d2_utils')
 const tsCsv = require('./timeseries_as_csv')
 const tsJson = require('./timeseries_as_json')
 
-const secInHour = 3600
-const milliseconds = 1000
-const hoursInDay = 24
-
-function deriveCsvFilePath (csvBasePath, type, timestamp, stationId, modelRun) {
-  // No test for this function
-  let formatString
-  let contentType
-  let extension
-  let subdir
-
-  if (type === 'REPORT') {
-    formatString = 'YYYYMMDD'
-    contentType = 'BEOB'
-    extension = '.csv'
-    subdir = 'poi'
-  } else if (type === 'MOSMIX_KMZ') {
-    formatString = 'YYYYMMDDHH'
-    contentType = 'MOSMIX'
-    extension = '.kmz'
-    subdir = 'mos'
-  } else {
-    formatString = 'YYYYMMDDHH'
-    contentType = 'MOSMIX'
-    extension = '.csv'
-    subdir = 'poi'
-  }
-
-  const dayDateTimeString =
-    moment
-      .utc(timestamp)
-      .format(formatString)
-      .slice(0, -2) + modelRun
-  const fileName = stationId + '-' + contentType + extension
-
-  return path.join(csvBasePath, subdir, dayDateTimeString, fileName)
-}
-
-async function getMosmixDataDirectories (
-  mosmixBasePath,
-  startTimestamp,
-  endTimestamp,
-  stationId,
-  modelRun
-) {
-  // No test for this function
-  const result = []
-  let timestamp = startTimestamp - secInHour * milliseconds
-  while (timestamp <= endTimestamp) {
-    // Timestamp is increased by 24 hours in iteration
-    // Model-run is explicitly += to the dayDateTimeString
-    const filePath = deriveCsvFilePath(
-      mosmixBasePath,
-      'MOSMIX_KMZ',
-      timestamp,
-      stationId,
-      modelRun
-    )
-    result.push(filePath)
-    timestamp += hoursInDay * secInHour * milliseconds
-  }
-  return [...new Set(result)]
-}
-
-async function readMosmixTimeseriesData (
-  mosmixBasePath,
-  voiConfigs,
-  startTimestamp,
-  endTimestamp,
-  stationId,
-  modelRun,
-  stationCatalog
-) {
-  // No test for this function
-  const resultArray = []
-  const mosmixFiles = await getMosmixDataDirectories(
-    mosmixBasePath,
-    startTimestamp,
-    endTimestamp,
-    stationId,
-    modelRun
-  )
-  for (let i = 0; i < mosmixFiles.length; i += 1) {
-    const filePath = mosmixFiles[i]
-    try {
-      const fileContent = await csv.extractKmlFile(filePath)
-      const result = await csv.parseKmlFile(fileContent)
-      resultArray.push(result)
-    } catch (error) {
-      log.warn(error, `failed to extract the .kml-file from ${filePath}`)
-    }
-  }
-  return resultArray
-}
 
 function dropTimeseriesDataNotOfInteresWithParameter (
   voiConfigs,
@@ -137,7 +36,7 @@ function getConfigFromModel (model, COSMO_DATA_BASE_PATH, MOSMIX_DATA_BASE_PATH)
     return {
       model: 'mosmix',
       allowedModelRuns: ['03', '09', '15', '21'],
-      functionToReadData: readMosmixTimeseriesData,
+      functionToReadData: bmu.readMosmixTimeseriesData,
       MODEL_DATA_PATH: MOSMIX_DATA_BASE_PATH,
       unitsConverter: gu.convertUnits,
       jsonRenderer: tsJson.renderTimeseriesAsJSON,
@@ -160,7 +59,6 @@ function getConfigFromModel (model, COSMO_DATA_BASE_PATH, MOSMIX_DATA_BASE_PATH)
   }
 }
 
-exports.readMosmixTimeseriesData = readMosmixTimeseriesData
 exports.dropTimeseriesDataNotOfInteresWithParameter = dropTimeseriesDataNotOfInteresWithParameter
 exports.checkIfValidModelRun = checkIfValidModelRun
 exports.getConfigFromModel = getConfigFromModel
